backend/pyproject.toml
[project]
name = "backend"
version = "1.0.0"
description = ""
authors = [
    {name = "gleb",email = "deffender1337@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqladmin (>=0.20.1,<0.21.0)",
    "fastapi[all] (>=0.115.11,<0.116.0)",
    "beanie (>=1.29.0,<2.0.0)",
    "motor (>=3.7.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
